nav bar working and grids are now homogenized except for map
