+++ client/src/components/NavBar/Nav.jsx
@@ -1,210 +1,73 @@
-import * as React from "react";
-import AppBar from "@mui/material/AppBar";
-import Box from "@mui/material/Box";
-import Toolbar from "@mui/material/Toolbar";
-import IconButton from "@mui/material/IconButton";
-import Typography from "@mui/material/Typography";
-import Menu from "@mui/material/Menu";
-// import MenuIcon from "@mui/icons-material/Menu";
-import Container from "@mui/material/Container";
-// import Avatar from "@mui/material/Avatar";
-import Button from "@mui/material/Button";
-// import Tooltip from "@mui/material/Tooltip";
-import MenuItem from "@mui/material/MenuItem";
+import React from "react";
+import { Header, Title, Image, Container } from "@mantine/core";
 import { Link } from "react-router-dom";
-// import useStyles from "./style"; // "./styles";
+import apharreA from "../../images/apharreA.png";
+import useStyles from "../../mantine/globalStyles";
 
-// import gopherRipped from "../../images/gopherRipped.png"; // "./images/gopherRipped.png";
-
-const pages = ["Map", "Posts", "Chart", "Contact Us"];
-// const settings = ["Profile", "Account", "Dashboard", "Logout"];
-
-function ResponsiveAppBar() {
-  const [anchorElNav, setAnchorElNav] = React.useState(null);
-  // const [anchorElUser, setAnchorElUser] = React.useState(null);
-  // const classes = useStyles();
-
-  const handleOpenNavMenu = (event) => {
-    setAnchorElNav(event.currentTarget);
-  };
-  // const handleOpenUserMenu = (event) => {
-  //   setAnchorElUser(event.currentTarget);
-  // };
-
-  const handleCloseNavMenu = () => {
-    setAnchorElNav(null);
-  };
-
-  // const handleCloseUserMenu = () => {
-  //   setAnchorElUser(null);
-  // };
+function NavBar() {
+  const { classes } = useStyles();
 
   return (
-    <AppBar position="static">
-      <Container maxWidth={false}>
-        <Toolbar disableGutters>
-          <Typography
-            variant="h6"
-            noWrap
-            component="div"
-            sx={{ mr: 2, display: { xs: "none", md: "flex" } }}
-          >
+    <Container size="max-width" px={10}>
+      <Header
+        height={70}
+        p="md"
+        backgroundColor="##3394e8"
+        className={classes.navBarHeader}
+      >
+        <div style={{ display: "flex", alignItems: "center", height: "100%" }}>
+          <Image height={65} src={apharreA} />
+          {/* <Title order={2} pl={20}> */}
+          <Title order={2} className={classes.navBarTitle}>
             Apharre
-          </Typography>
-
-          <Box sx={{ flexGrow: 1, display: { xs: "flex", md: "none" } }}>
-            <IconButton
-              size="large"
-              aria-label="account of current user"
-              aria-controls="menu-appbar"
-              aria-haspopup="true"
-              onClick={handleOpenNavMenu}
-              color="inherit"
-            >
-              {/* <MenuIcon /> */}
-            </IconButton>
-            <Menu
-              id="menu-appbar"
-              anchorEl={anchorElNav}
-              anchorOrigin={{
-                vertical: "bottom",
-                horizontal: "left",
-              }}
-              keepMounted
-              transformOrigin={{
-                vertical: "top",
-                horizontal: "left",
-              }}
-              open={Boolean(anchorElNav)}
-              onClose={handleCloseNavMenu}
-              sx={{
-                display: { xs: "block", md: "none" },
-              }}
-            >
-              {pages.map((page) => (
-                <MenuItem key={page} onClick={handleCloseNavMenu}>
-                  <Typography textAlign="center">
-                    <Link to={`/${page}`}>{page}</Link>
-                  </Typography>
-                </MenuItem>
-              ))}
-            </Menu>
-          </Box>
-          <Typography
-            variant="h6"
-            noWrap
-            component="div"
-            sx={{ flexGrow: 1, display: { xs: "flex", md: "none" } }}
+          </Title>
+          <Title
+            component={Link}
+            variant="link"
+            to="/map"
+            order={3}
+            className={classes.navBarTitle}
           >
-            LOGO
-          </Typography>
-          <Box sx={{ flexGrow: 1, display: { xs: "none", md: "flex" } }}>
-            {pages.map((page) => (
-              <Button
-                key={page}
-                onClick={handleCloseNavMenu}
-                sx={{ my: 2, color: "white", display: "block" }}
-              >
-                <Link
-                  style={{ textDecoration: "none", color: "black" }}
-                  to={`/${page}`}
-                >
-                  {page}
-                </Link>
-              </Button>
-            ))}
-          </Box>
-          <Box sx={{ flexGrow: 0 }}>
-            {/* <img
-              // className={classes.image}
-              src={gopherRipped}
-              alt="memories"
-              height="60"
-            /> */}
-          </Box>
-          {/* <Box sx={{ flexGrow: 0 }}>
-            <Tooltip title="Open settings">
-              <IconButton onClick={handleOpenUserMenu} sx={{ p: 0 }}>
-                <Avatar alt="Apharre" src="/static/images/avatar/2.jpg" />
-              </IconButton>
-            </Tooltip>
-            <Menu
-              sx={{ mt: "45px" }}
-              id="menu-appbar"
-              anchorEl={anchorElUser}
-              anchorOrigin={{
-                vertical: "top",
-                horizontal: "right",
-              }}
-              keepMounted
-              transformOrigin={{
-                vertical: "top",
-                horizontal: "right",
-              }}
-              open={Boolean(anchorElUser)}
-              onClose={handleCloseUserMenu}
-            >
-              {settings.map((setting) => (
-                <MenuItem key={setting} onClick={handleCloseUserMenu}>
-                  <Typography textAlign="center">{setting}</Typography>
-                </MenuItem>
-              ))}
-            </Menu>
-          </Box> */}
-        </Toolbar>
-      </Container>
-    </AppBar>
+            Map
+          </Title>
+          <Title
+            component={Link}
+            variant="link"
+            to="/chart"
+            order={3}
+            className={classes.navBarTitle}
+          >
+            Chart
+          </Title>
+          <Title
+            component={Link}
+            variant="link"
+            to="/contact"
+            order={3}
+            className={classes.navBarTitle}
+          >
+            Contact Us
+          </Title>
+          <Title
+            component={Link}
+            variant="link"
+            to="/posts"
+            order={3}
+            className={classes.navBarTitle}
+          >
+            Posts
+          </Title>
+        </div>
+
+        {/* <Routes>
+          <Route path="/map" element={<Map />} />
+          <Route path="/chart" element={<ChartPage />} />
+          <Route path="/contact" element={<Map />} />
+          <Route path="/posts" element={<PostPage />} />
+        </Routes> */}
+      </Header>
+    </Container>
   );
 }
 
-export default ResponsiveAppBar;
-// import React from "react";
-// import { BrowserRouter as Router, Routes, Route, Link } from "react-router-dom";
-
-// function Home() {
-//   return <h2>Home</h2>;
-// }
-
-// function About() {
-//   return <h2>About</h2>;
-// }
-
-// function Users() {
-//   return <h2>Users</h2>;
-// }
-
-// export default function BasicRoute() {
-//   return (
-//     <Router>
-//       <div>
-//         <nav>
-//           <ul>
-//             <li>
-//               <Link to="/">Home</Link>
-//             </li>
-//             <li>
-//               <Link to="/about">About</Link>
-//             </li>
-//             <li>
-//               <Link to="/users">Users</Link>
-//             </li>
-//           </ul>
-//         </nav>
-
-//         {/* A <Switch> looks through its children <Route>s and
-//             renders the first one that matches the current URL. */}
-//         <Routes>
-//           <Route path="/about">
-//             <About />
-//           </Route>
-//           <Route path="/users">
-//             <Users />
-//           </Route>
-//           <Route path="/">
-//             <Home />
-//           </Route>
-//         </Routes>
-//       </div>
-//     </Router>
-//   );
-// }
+export default NavBar;
